fix: move code from photo-modal.js to fullsize.js, delete file photo-modal.js
