+++ js/main.js
@@ -1,6 +1,6 @@
 import { getPhotosDataArray } from './data.js';
 import { renderThumbnails } from './thumbnails.js';
-import './photo-modal.js';
+import './fullsize.js';
 import './upload-form.js';
 
 getPhotosDataArray();
