--- js/photo-modal.js
@@ -1,40 +1,0 @@
-import { isEscapeKey } from './util.js';
-
-const bigPictureModal = document.querySelector('.big-picture');
-const closeModalButton = document.querySelector('.big-picture__cancel');
-const loaderButton = document.querySelector('.social__comments-loader');
-
-const onDocumentEscKeyDown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closePhotoModal();
-  }
-};
-
-function openPhotoModal() {
-  bigPictureModal.classList.remove('hidden');
-  document.body.classList.add('modal-open');
-
-  document.addEventListener('keydown', onDocumentEscKeyDown);
-}
-
-function closePhotoModal() {
-  bigPictureModal.classList.add('hidden');
-  document.body.classList.remove('modal-open');
-
-  const commentsListArray = document.querySelectorAll('.social__comment');
-  commentsListArray.forEach((element) => element.remove());
-
-  loaderButton.classList.remove('hidden');
-
-  document.removeEventListener('keydown', onDocumentEscKeyDown);
-}
-
-// закрытие фото в модальном окне
-closeModalButton.addEventListener('click', () => {
-  closePhotoModal();
-});
-
-export { openPhotoModal };
-
-
